Refactor : separate dao & services
