+++ src/main/java/app/rl/blog/dao/PostDAO.java
@@ -1,13 +1,14 @@
-package app.rl.blog.service;
+package app.rl.blog.dao;
 
 import java.util.List;
 
 import app.rl.blog.entity.Post;
 import jakarta.validation.Valid;
 
-public interface PostService {
+public interface PostDAO {
 
     public Post savePost(@Valid Post post);
     public List<Post> fetchPostList();
+    public List<Post> fetchPostsByCategoryId(Long id);
     public void deletePostById(Long id);
 }
