+++ src/main/java/app/rl/blog/dao/PostDAOImpl.java
@@ -1,4 +1,4 @@
-package app.rl.blog.service;
+package app.rl.blog.dao;
 
 import java.util.List;
 
@@ -7,12 +7,15 @@ import org.springframework.stereotype.Service;
 
 import app.rl.blog.entity.Post;
 import app.rl.blog.repository.PostRepository;
+import jakarta.persistence.EntityManager;
+import jakarta.persistence.TypedQuery;
 
 @Service
-public class PostServiceImpl implements PostService {
-    
+public class PostDAOImpl implements PostDAO {
+
     @Autowired
     public PostRepository postRepository;
+    private EntityManager em;
 
     @Override
     public Post savePost(Post post) {
@@ -30,5 +33,16 @@ public class PostServiceImpl implements PostService {
     public void deletePostById(Long id) {
 
         postRepository.deleteById(id);
+    }
+
+    public List<Post> fetchPostsByCategoryId(Long id) {
+
+        TypedQuery<Post> query = em.createQuery(
+                "from Post where category.id = :data",
+                Post.class);
+
+        query.setParameter("data", id);
+
+        return query.getResultList();
     }
 }
